requirements.txt
fastapi
uvicorn
Pillow
google-cloud-storage
requests